add Training activity and class | changed activity_main
