+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/Acer/Documents/AndroidProjects/BrainTraining/app/src/main/res/layout/activity_main.xml" value="0.22010869565217392" />
+        <entry key="..\:/Users/Acer/Documents/AndroidProjects/BrainTraining/app/src/main/res/layout/activity_training.xml" value="0.22447916666666667" />
       </map>
     </option>
   </component>
